+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 
 Bugfixes:
   - LokiJS `findDocumentsById()` returned additional `$loki` property.
+  - LokiJS `bulkAddRevisions()` must not mutate the input.
 
 ### 10.3.4 (7 November 2021)